+++ CHANGELOG.md
@@ -1,11 +1,29 @@
-# 0.0.5
+# Changelog
 
-## Added
+## [Unreleased] - ReleaseDate
+
+### Added
+* support for setting tunable jail parameters
+* support for non-persistent jails
+
+### Changed
+* examples/jls: fixed `nix` version mismatch with `rctl` crate
+* fixed jail teardown and save if RCTL not enabled
+* `RunningJail` now derives `Copy`.
+* `RunningJail::jail_attach` is now public.
+* `RunningJail::save` now no longer saves the `vnet` parameter if it is set to
+  `inherit` (2). See #34.
+* updated rctl to 0.0.5
+* updated prettytable-rs to 0.8.0
+
+## [0.0.5] - 2018-07-05
+
+### Added
 * RCTL / RACCT support
 
-# 0.0.4
+## [0.0.4] - 2018-06-21
 
-## Added
+### Added
 * this Changelog
 
 * iteration over running jails